Add TelemetryVue application to allow visualization of Telemetry Datasources
TelemetryVue is broken down into 3 components:
- Model, which knows nothing about Swing and is used to track and update
  state
- GUI, which interfaces the Model with Swing, providing events from the
  user back to the model, and reacting to events from the model
- H2, which is the H2-specific database opening logic

Unit tests on the UI are not extensive at this point, however it was
designed to make the model and H2 components easily testable.

In theory it is also possible to provide mock input to the UI and
provide a mock model implementation which would make testing the UI more
possible.

There are still many possible future enhancements for the data model and
application UI, but this provides a minimum-viable-product to allow
tabulating data from a database.